requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3

torch==2.9.1
torchvision==0.24.1
numpy==2.1.3
pillow==11.0.0

mlflow==2.19.0

faiss-cpu==1.12.0
sentence-transformers==3.3.1

pytest==8.3.4


ragas==0.2.7
datasets==2.21.0
langchain-openai==0.2.14
